+++ client/css/bem-components/banners.less
