+++ client/css/bem-components/buttons.less
@@ -1,5 +1,3 @@
-/* Global REM Conversion 1rem = 16px */
-
 .button {
   background: @buttonBackground;
   background-image: @buttonGradient;
