+++ client/css/bem-components/empty-view.less
