+++ client/css/bem-components/icons-table.less
