+++ client/css/bem-components/json-view.less
