+++ client/css/bem-components/paginator.less
