+++ client/css/bem-components/properties-table.less
