+++ client/css/bem-components/query-history.less
