+++ client/css/bem-components/section-banner.less
