+++ client/css/bem-components/section-controls.less
