+++ client/css/bem-components/section-header.less
