+++ client/css/bem-components/section-subtabs.less
