+++ client/css/bem-components/step-indicator.less
