+++ client/css/bem-components/tabs.less
