+++ client/css/bem-components/view-modal.less
