+++ client/css/global/fonts.less
@@ -1,0 +1,72 @@
+
+/* Fonts! */
+//Reg
+@font-face {
+    font-family: 'Akzidenz';
+    src: url('../../assets/font/akzidgrostdreg.eot');
+    src: url('../../assets/font/akzidgrostdreg.eot#iefix') format('embedded-opentype'),
+    url('../../assets/font/akzidgrostdreg.woff') format('woff'),
+    url('../../assets/font/akzidgrostdreg.ttf')  format('truetype');
+    font-style: normal;
+    font-weight: normal;
+}
+//Reg Italic
+@font-face {
+    font-family: 'Akzidenz';
+    src: url('../../assets/font/akzidgrostdita.eot');
+    src: url('../../assets/font/akzidgrostdita.eot#iefix') format('embedded-opentype'),
+    url('../../assets/font/akzidgrostdita.woff') format('woff'),
+    url('../../assets/font/akzidgrostdita.ttf')  format('truetype');
+    font-style: italic;
+    font-weight: normal;
+}
+//Bold
+@font-face {
+    font-family: 'Akzidenz';
+    src: url('../../assets/font/akzidgrostdmed.eot');
+    src: url('../../assets/font/akzidgrostdmed.eot#iefix') format('embedded-opentype'),
+    url('../../assets/font/akzidgrostdmed.woff') format('woff'),
+    url('../../assets/font/akzidgrostdmed.ttf')  format('truetype');
+    font-style: normal;
+    font-weight: bold;
+}
+//Bold Italic
+@font-face {
+    font-family: 'Akzidenz';
+    src: url('../../assets/font/akzidgrostdmedita.eot');
+    src: url('../../assets/font/akzidgrostdmedita.eot#iefix') format('embedded-opentype'),
+    url('../../assets/font/akzidgrostdmedita.woff') format('woff'),
+    url('../../assets/font/akzidgrostdmedita.ttf')  format('truetype');
+    font-style: italic;
+    font-weight: bold;
+}
+//Light
+@font-face {
+    font-family: 'Akzidenz';
+    src: url('../../assets/font/akzidgrostdlig.eot');
+    src: url('../../assets/font/akzidgrostdlig.eot#iefix') format('embedded-opentype'),
+    url('../../assets/font/akzidgrostdlig.woff') format('woff'),
+    url('../../assets/font/akzidgrostdlig.ttf')  format('truetype');
+    font-style: normal;
+    font-weight: 200;
+}
+//Light Italic
+@font-face {
+    font-family: 'Akzidenz';
+    src: url('../../assets/font/akzidgrostdligita.eot');
+    src: url('../../assets/font/akzidgrostdligita.eot#iefix') format('embedded-opentype'),
+    url('../../assets/font/akzidgrostdligita.woff') format('woff'),
+    url('../../assets/font/akzidgrostdligita.ttf')  format('truetype');
+    font-style: italic;
+    font-weight: 200;
+}
+
+@font-face {
+    font-family: 'MMS Icons';
+    font-style: normal;
+    font-weight: normal;
+    src: url('../../assets/font/MMSIcons-Regular.eot'); /* IE9 Compat Modes */
+    src: url('../../assets/font/MMSIcons-Regular.eot#iefix') format('embedded-opentype'),
+    url('../../assets/font/MMSIcons-Regular.woff') format('woff'),
+    url('../../assets/font/MMSIcons-Regular.ttf')  format('truetype'); /* Safari, Android, iOS */
+}
