+++ client/css/global/mms-icons.less
