+++ client/css/root.less
@@ -1,48 +1,49 @@
 // @import 'normalize.css';
 // Global
-@import 'fonts.less';
-@import 'typography.less';
-@import 'grid.less';
-@import 'mms-icons.less';
-@import 'palette.less';
-@import 'utilities.less';
+@import 'global/fonts.less';
+@import 'global/typography.less';
+@import 'global/grid.less';
+@import 'global/mms-icons.less';
+@import 'global/palette.less';
+@import 'global/colors.less';
+@import 'global/utilities.less';
 
 // BEM Components
-@import 'alert.less';
-@import 'bubble.less';
-@import 'buttons.less';
-@import 'button-group.less';
-@import 'banners.less';
-@import 'code.less';
-@import 'colors.less'; // to be removed
-@import 'forms.less';
-@import 'rich-radio.less';
-@import 'tabs.less';
-@import 'plain-table.less';
-@import 'tables.less';
-@import 'json-view.less';
-@import 'view-modal.less';
-@import 'checklist.less';
-@import 'nested-list.less';
-@import 'number-circle.less';
-@import 'step-indicator.less';
-@import 'paginator.less';
-@import 'home.less';
-@import 'navigation.less';
-@import 'empty-view.less';
-@import 'properties-table.less';
-@import 'data-viz-row.less';
-@import 'summary-table.less';
-@import 'query-history.less';
-@import 'mms-icons.less';
-@import 'icons-table.less';
-@import 'section-header.less';
-@import 'section-banner.less';
-@import 'section-controls.less';
-@import 'section-subtabs.less';
-@import 'section-warnings.less';
-@import 'guidelines-modify.less';
-@import 'checkbox.less';
+@import 'bem-components/alert.less';
+@import 'bem-components/bubble.less';
+@import 'bem-components/buttons.less';
+@import 'bem-components/button-group.less';
+@import 'bem-components/banners.less';
+@import 'bem-components/code.less';
+@import 'bem-components/forms.less';
+@import 'bem-components/rich-radio.less';
+@import 'bem-components/tabs.less';
+@import 'bem-components/plain-table.less';
+@import 'bem-components/tables.less';
+@import 'bem-components/json-view.less';
+@import 'bem-components/view-modal.less';
+@import 'bem-components/checklist.less';
+@import 'bem-components/nested-list.less';
+@import 'bem-components/number-circle.less';
+@import 'bem-components/step-indicator.less';
+@import 'bem-components/paginator.less';
+@import 'bem-components/navigation.less';
+@import 'bem-components/empty-view.less';
+@import 'bem-components/properties-table.less';
+@import 'bem-components/data-viz-row.less';
+@import 'bem-components/summary-table.less';
+@import 'bem-components/query-history.less';
+@import 'bem-components/icons-table.less';
+@import 'bem-components/section-header.less';
+@import 'bem-components/section-banner.less';
+@import 'bem-components/section-controls.less';
+@import 'bem-components/section-subtabs.less';
+@import 'bem-components/section-warnings.less';
+@import 'bem-components/checkbox.less';
+
+// Site Styling
+@import 'site/home.less';
+@import 'site/guidelines-modify.less';
 
 
 * {
